fix: rename auth extensions
